Initial Commit
Fixed some errors in R2_custom enough to get it running, functionality
should be fine but haven't checked. Started work on the Extended Health
Clinic model, adding entities and that sort of thing. Controller logic
and some other miscellaneous things still need to be done.

On the errors in R2_custom: there seems to be some conflict between
DisplayEntity and Entity, where the library is trying to use
DisplayEntity in super methods derived from the parent class, but Java
won't allow you to change the return type, even if DisplayEntity is a
subclass of Entity. Not sure what the best fix is, for now I've just
changed DisplayEntity to Entity where necessary.