More Progress and Bug Fixes
Fixed bugs related to entities getting stuck in the waiting room if
their next activity was not free for any reason. Now, if none of the
if-else options are satisfied, a new activity signal will be sent 5
seconds later to run through the controller logic again, and this repeats
until the entity moves on. This time can be changed with the
retrywaittime variable, which is a property of the controller. Fixed
various bugs related to the logic governing entity flow, preventing
entities from getting stuck. Fixed bugs related to entities not having
the correct availability behaviour.

EHC model is now fully functioning. Still to do: fixing statistics
objects, validation of output compared to data provided for the
tutorial.