Travel Fix and Other Improvements
Added travel times for patients between activities using entity conveyors. Added groundwork for end of day/end of sim events to wind down and allow doctors and nurses to exit. Added a max queue length property to the HCCM Control Activity class as well as setter and getter methods for this property.