Finishing Off Travel and Utilization Statistics
Changed the functionality of maxQueueLength to be a key input rather
than an internal property. This makes it easier for the user to edit,
but keeps the same functionality. Switched out the gate and queue for a
branch that directs incoming patients to either the waiting room or the
leave conveyor. Added a custom output to the controller to allow it to
control the function of the branch. Added basic end of sim (192h) logic
to allow the doctors/nurses to exit. Not strictly necessary to collect
utilization data but does make it easier to see and understand. Fixed
exiting of nurses (was direct to sink rather than through conveyor and
stats). Fixed new logic that replaced the recursive/looping activity
creation resulting in much faster simulation times.

A note on statistics: the statistical output still shows descrepancies
compared with the twiki EHC results. In particular:
- Walkup.WaitTriage is roughly double (0.131 vs 0.0668)
- Walkup.WaitTest is closer but still off (0.00257 vs 0.00195)
- Walkup.WaitTreat is more than double (0.435 vs 0.182)
- Appointment.WaitTreat is much larger (0.228 vs 0.129)
- (Based on this you can imagine the differences in overall wait and
  system times)
- Utilizations are not that far off (52.4% vs 50.0% and 80.5% vs 84.3%)

Possible explanations: in this model the patients cannot travel along
the conveyors until the server at the end of it is free. I believe in
the twiki model they can travel along the conveyor and queue at its end,
so this could cause a significant results discrepancy.