Created new model for EHC with Rosters
Due to needing to have a period of no patients in order to safely switch
the doctors between the walk up and treatment rooms, I added some
basic start and end of day functionality to allow for this. This will
prevent validation of the outputs but since the core functionality is
identical to the regular EHC I don't see it as a huge issue. Regardless,
this can fairly easily be edited to work in a variety of ways that all
implement the rostering functionality, so anyone can feel free to do so.