Trying Netbeans: edits to model preparing for fix
Changes:

- added some things to the .gitignore that are Netbeans specific
- Added stats objects with completed inputs
- Added branches for after triage with completed inputs
- Added assign object to set needsTest attribute for walkup patients
- Removed log events and TriageEndTrigger

Model should be basically complete now, waiting on secondary fix to
allow for non hccm objects to be entered as inputs for the
nextActivityEventList attribute for process activities/wait activities.

On Netbeans: switched computers and was struggling to get eclipse
running without jogl/display errors, so in the mean time will be using
Netbeans. Hopefully no compatibility issues as far as code sharing goes.