Changes to ProcessActivity.java and model updates
Made some changes to sendToNextComponent() to try and fix issues with
entitycontainers vs displayentities. Seems to be mostly functional now
although there are some visual glitches that I'm hoping can be fixed in
finish(). Will see what Mike thinks of these changes, they may be
reversed/adapted, it does seem like our functionalities are differing
even when we should be using the same files so not sure what is going on
there.

Model updates are simple changes to nextAEJ and adding a branch to
organize the doctors after treatment so they return to the proper place
(fixing a bug Mike noted).