Changes to jack_ehc.cfg and triggers
Fixed triggers and made some model changes. Model is now almost
functional. There is an issue when an entity is sent directly from a branch to a wait activity, as in the case of the triage nurse going from PatientOrNurseBranch1 to WaitToTriage, it doesn’t generate a new request as it would if they came from a ProcessActivity. So obviously the model stalls after the first triage.