Changes to cfg and triggers
Based on Mike's feedback that only the customer should send a request
and the servers should just fulfil those based on their convenience. Not
currently functioning, there are issues when patients arrive while the
server is servicing that are yet to be resolved, plus likely some
additional bugs that will get fixed afterwards.