Changes to model parameters for validation
Various changes to model parameters to match the regular jaamsim model.
Output is getting closer now, some examples:

WalkUpWaitForTriage: 0.0665 vs 0.0688h
WalkUpWaitForTest: 0.00604 vs 0.00195h
WalkUpWaitForTreat: 0.209 vs 0.182h
ScheduledWaitForTreat: 0.132 vs 0.129h

where the original model results are second. This is already better than
the rick and rick model (although that does have some changes to the
model and may also have similar problems with model parameters that have
not been found yet). WalkUpWaitForTreat is a little high but probably
acceptable, WaitForTest being high is more of a concern, could be a
conflict with test times (using set 10 min vs distribution fixed at 10
mins) otherwise not sure. There shouldn't be many possible sources of an
error in that area, the test part of the model is small and
self-contained so hopefully it is something obvious.