Added convenience functions
Added AttributeCompare to sort a list of entities based on the value of and attribute.
Added getEntitiesInState to get entities of a given type in a given state, useful to get all waiting entites.
Added getSubModelEntity use to refer directly to components that could be in a submodel but not always